// frontend/src/lib/api/models/OrganizerInputDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OrganizerInputDto = {
    /**
     * User ID (UUID v4)
     */
    userId?: string;
    /**
     * Free-text organizer name when not a system user
     */
    displayName?: string;
};